ftx: fix ftx test cases

+++ pkg/exchange/ftx/ftxapi/trade.go
@@ -120,6 +120,7 @@ func (c *RestClient) NewCancelOrderByClientOrderIdRequest(clientOrderId string) 
 }
 
 type Fill struct {
+	// Id is fill ID
 	Id            uint64           `json:"id"`
 	Future        string           `json:"future"`
 	Liquidity     Liquidity        `json:"liquidity"`
@@ -127,7 +128,7 @@ type Fill struct {
 	BaseCurrency  string           `json:"baseCurrency"`
 	QuoteCurrency string           `json:"quoteCurrency"`
 	OrderId       uint64           `json:"orderId"`
-	TradeId       int              `json:"tradeId"`
+	TradeId       uint64             `json:"tradeId"`
 	Price         fixedpoint.Value `json:"price"`
 	Side          Side             `json:"side"`
 	Size          fixedpoint.Value `json:"size"`
